Factor UrlProviderService out of ModuleLoader
