+++ api/src/org/labkey/api/action/UrlProviderOverrideHandler.java
@@ -2,7 +2,6 @@ package org.labkey.api.action;
 
 import org.labkey.api.module.Module;
 import org.labkey.api.module.ModuleDependencySorter;
-import org.labkey.api.module.ModuleLoader;
 import org.labkey.api.util.Pair;
 
 import java.lang.reflect.InvocationHandler;
@@ -29,7 +28,7 @@ public class UrlProviderOverrideHandler implements InvocationHandler
     @Override
     public Object invoke(Object proxy, Method m, Object[] args) throws Throwable {
         // check to see if the given interface has any module registered overrides
-        List<Pair<Module, Class<? extends UrlProvider>>> impls = ModuleLoader.getInstance().getUrlProviderOverrides(_inter);
+        List<Pair<Module, Class<? extends UrlProvider>>> impls = UrlProviderService.getInstance().getUrlProviderOverrides(_inter);
         if (impls != null)
         {
             // convert the pairs of module/impl to a map from module -> impl and a list of all module dependencies
